[#8] refactor: componentScan 을 위한 패키지 이동
